Fixed a harmless error message when rewrite-availability-macros.sh is run in a clean build.
Rubber-stamped by Anders Carlsson.

* mac/rewrite-availability-macros.sh:



Canonical link: https://commits.webkit.org/151785@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@169905 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/mac/rewrite-availability-macros.sh
@@ -30,7 +30,7 @@ fi
 
 TIMESTAMP_PATH=${TARGET_TEMP_DIR}/${0##*/}
 
-if [[ $0 -nt $TIMESTAMP_PATH ]]; then
+if [[ -e $TIMESTAMP_PATH && $0 -nt $TIMESTAMP_PATH ]]; then
     rm "${TIMESTAMP_PATH}";
 fi
